Move SkillsBench post-run debug gate inward
